Free_Data_API-feat-microservice-refactor/requirements.txt
# 核心套件
pandas
duckdb
pyyaml
python-dotenv
requests
requests-cache
